FX-625 - check for valid ACL assignments, don't remove first and required ACL assignment
git-svn-id: http://svn.flexive.org/repository/flexive/trunk@1530 d35d4baa-7912-0410-b0ae-edacad209e15

+++ src/framework/java/com/flexive/shared/security/ACL.java
@@ -47,7 +47,7 @@ public class ACL extends AbstractSelectableObjectWithName implements Serializabl
     /**
      * ID for an empty placeholder ACL.
      */
-    public static final int NULL_ACL_ID = 0;
+    public static final long NULL_ACL_ID = 0;
 
     /**
      * Max. id for internal ACL's (needed for internal checks)
